only fail fast in browser tests, not for qawolf tests
- split out runTest from Browser and retryAsync and $xText to pageUtils
