+++ src/jest.failFast.ts
